CreateProcess and movements around it.

+++ py2http/tests/utils_for_testing.py
@@ -1,27 +1,25 @@
 from contextlib import contextmanager
+from multiprocessing.context import Process
+from time import sleep
+from py2http.util import conditional_logger, deprecate
 
-def conditional_logger(verbose=False):
-    if verbose:
-        clog = print
-    else:
-        def clog(*args, **kwargs):
-            pass  # do nothing
-    return clog
 
-@contextmanager  # see https://docs.python.org/3/library/contextlib.html#contextlib.contextmanager
-def run_server(launcher, wait_before_entering=0.5, verbose=False, **kwargs):
+@contextmanager
+@deprecate
+def run_server(launcher, wait_before_entering=0, verbose=False, **kwargs):
     """Context manager to launch server on entry, and shut it down on exit"""
-    from multiprocessing.context import Process
-    from time import sleep
+    from warnings import warn
     clog = conditional_logger(verbose)
     server = None
     try:
         server = Process(target=launcher, kwargs=kwargs)
+        clog(f"Starting server...")
         server.start()
-        clog(f"Server starting")
+        clog(f"... server started.")
         sleep(wait_before_entering)
         yield server
     finally:
-        if server is not None:
+        if server is not None and server.is_alive():
+            clog(f"Terminating server...")
             server.terminate()
-        clog(f"Server stopped")
+        clog(f"... server terminated")
